+++ docs/CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.2a8 (unreleased)
 ------------------
 
+- Fix ``test_warning_for_uneditable_content`` to work with recent browser layer
+  changes in ``plone.app.z3cform``.
+  [thet]
+
 - Update image_view_fullscreen.pt for mobile friendliness.
   [fulv]